Split and link `ftkirbyspecialpeach` and `ftkirbyspecialgamewatch`
Also split `ftkirbyspecialmars`
