+++ config/GALE01/symbols.txt
@@ -23708,7 +23708,8 @@ it_803F5468 = .data:0x803F5468; // type:object size:0x70 scope:global
 it_803F54D8 = .data:0x803F54D8; // type:object size:0xD0 scope:global
 @285 = .data:0x803F55A8; // type:object size:0x25 scope:local data:string
 it_803F55D0 = .data:0x803F55D0; // type:object size:0xC0 scope:global
-it_803F5690 = .data:0x803F5690; // type:object size:0x25 scope:local data:string
+...data.0 = .data:0x803F55D0; // type:label scope:local
+@243 = .data:0x803F5690; // type:object size:0x25 scope:local data:string
 it_803F56B8 = .data:0x803F56B8; // type:object size:0x10 scope:global
 it_803F56C8 = .data:0x803F56C8; // type:object size:0x50 scope:global
 @234 = .data:0x803F5718; // type:object size:0x25 scope:local data:string
@@ -27179,8 +27180,8 @@ it_804D51C8 = .sdata:0x804D51C8; // type:object size:0x8 scope:global
 @377 = .sdata:0x804D51E8; // type:object size:0x5 scope:local data:string
 @231 = .sdata:0x804D51F0; // type:object size:0x7 scope:local data:string
 @232 = .sdata:0x804D51F8; // type:object size:0x5 scope:local data:string
-it_804D5200 = .sdata:0x804D5200; // type:object size:0x7 scope:global data:string
-it_804D5208 = .sdata:0x804D5208; // type:object size:0x5 scope:global data:string
+@241 = .sdata:0x804D5200; // type:object size:0x7 scope:local data:string
+@242 = .sdata:0x804D5208; // type:object size:0x5 scope:local data:string
 @188 = .sdata:0x804D5210; // type:object size:0x7 scope:local data:string
 @189 = .sdata:0x804D5218; // type:object size:0x5 scope:local data:string
 @236 = .sdata:0x804D5220; // type:object size:0x7 scope:local data:string
@@ -30297,15 +30298,15 @@ ftKb_Init_804D956C = .sdata2:0x804D956C; // type:object size:0x4 scope:global al
 ftKb_Init_804D9570 = .sdata2:0x804D9570; // type:object size:0x4 scope:global align:4 data:float
 ftKb_Init_804D9574 = .sdata2:0x804D9574; // type:object size:0x4 scope:global align:4 data:float
 ftKb_Init_804D9578 = .sdata2:0x804D9578; // type:object size:0x8 scope:global align:8 data:double
-ftKb_Init_804D9580 = .sdata2:0x804D9580; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D9584 = .sdata2:0x804D9584; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D9588 = .sdata2:0x804D9588; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D958C = .sdata2:0x804D958C; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D9590 = .sdata2:0x804D9590; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D9594 = .sdata2:0x804D9594; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D9598 = .sdata2:0x804D9598; // type:object size:0x4 scope:global align:4 data:float
-ftKb_Init_804D95A0 = .sdata2:0x804D95A0; // type:object size:0x8 scope:global align:8 data:double
-ftKb_Init_804D95A8 = .sdata2:0x804D95A8; // type:object size:0x4 scope:global align:4 data:float
+@237 = .sdata2:0x804D9580; // type:object size:0x4 scope:local align:4 data:float
+@238 = .sdata2:0x804D9584; // type:object size:0x4 scope:local align:4 data:float
+@276 = .sdata2:0x804D9588; // type:object size:0x4 scope:local align:4 data:float
+@444 = .sdata2:0x804D958C; // type:object size:0x4 scope:local align:4 data:float
+@247 = .sdata2:0x804D9590; // type:object size:0x4 scope:local align:4 data:float
+@248 = .sdata2:0x804D9594; // type:object size:0x4 scope:local align:4 data:float
+@249 = .sdata2:0x804D9598; // type:object size:0x4 scope:local align:4 data:float
+@251 = .sdata2:0x804D95A0; // type:object size:0x8 scope:local align:8 data:double
+@287 = .sdata2:0x804D95A8; // type:object size:0x4 scope:local align:4 data:float
 @273 = .sdata2:0x804D95B0; // type:object size:0x4 scope:local align:4 data:float
 @276 = .sdata2:0x804D95B4; // type:object size:0x4 scope:local align:4 data:float
 @295 = .sdata2:0x804D95B8; // type:object size:0x4 scope:local align:4 data:float
@@ -32968,15 +32969,15 @@ it_804DC704 = .sdata2:0x804DC704; // type:object size:0x4 scope:global align:4 d
 @1121 = .sdata2:0x804DC860; // type:object size:0x8 scope:local align:8 data:double
 @1588 = .sdata2:0x804DC868; // type:object size:0x4 scope:local align:4 data:float
 @1589 = .sdata2:0x804DC86C; // type:object size:0x4 scope:local align:4 data:float
-it_804DC870 = .sdata2:0x804DC870; // type:object size:0x4 scope:global align:4 data:float
-it_804DC874 = .sdata2:0x804DC874; // type:object size:0x4 scope:global align:4 data:float
-it_804DC878 = .sdata2:0x804DC878; // type:object size:0x4 scope:global align:4 data:float
-it_804DC880 = .sdata2:0x804DC880; // type:object size:0x8 scope:global align:8 data:double
-it_804DC888 = .sdata2:0x804DC888; // type:object size:0x4 scope:global align:4 data:float
-it_804DC88C = .sdata2:0x804DC88C; // type:object size:0x4 scope:global align:4 data:float
-it_804DC890 = .sdata2:0x804DC890; // type:object size:0x8 scope:global align:8 data:double
-it_804DC898 = .sdata2:0x804DC898; // type:object size:0x8 scope:global align:8 data:double
-it_804DC8A0 = .sdata2:0x804DC8A0; // type:object size:0x4 scope:global align:4 data:float
+@188 = .sdata2:0x804DC870; // type:object size:0x4 scope:local align:4 data:float
+@189 = .sdata2:0x804DC874; // type:object size:0x4 scope:local align:4 data:float
+@190 = .sdata2:0x804DC878; // type:object size:0x4 scope:local align:4 data:float
+@191 = .sdata2:0x804DC880; // type:object size:0x8 scope:local align:8 data:double
+@192 = .sdata2:0x804DC888; // type:object size:0x4 scope:local align:4 data:float
+@193 = .sdata2:0x804DC88C; // type:object size:0x4 scope:local align:4 data:float
+@194 = .sdata2:0x804DC890; // type:object size:0x8 scope:local align:8 data:double
+@195 = .sdata2:0x804DC898; // type:object size:0x8 scope:local align:8 data:double
+@196 = .sdata2:0x804DC8A0; // type:object size:0x4 scope:local align:4 data:float
 @156 = .sdata2:0x804DC8A8; // type:object size:0x4 scope:local align:4 data:float
 @157 = .sdata2:0x804DC8AC; // type:object size:0x4 scope:local align:4 data:float
 @176 = .sdata2:0x804DC8B0; // type:object size:0x4 scope:local align:4 data:float
@@ -34187,8 +34188,8 @@ MSL_Math_804DE1B0 = .sdata2:0x804DE1B0; // type:object size:0x4 scope:global dat
 MSL_Math_804DE1B4 = .sdata2:0x804DE1B4; // type:object size:0x4 scope:global data:4byte
 __float_nan = .sdata2:0x804DE1B8; // type:object size:0x4 scope:global align:4 data:float
 __float_huge = .sdata2:0x804DE1BC; // type:object size:0x4 scope:global align:4 data:float
-MSL_Math_804DE1C0 = .sdata2:0x804DE1C0; // type:object size:0x4 scope:global align:4 data:float
-MSL_Math_804DE1C8 = .sdata2:0x804DE1C8; // type:object size:0x8 scope:global align:8 data:double
+@40 = .sdata2:0x804DE1C0; // type:object size:0x4 scope:local align:4 data:float
+@42 = .sdata2:0x804DE1C8; // type:object size:0x8 scope:local align:8 data:double
 THPDec_804DE1D0 = .sdata2:0x804DE1D0; // type:object size:0x4 scope:local data:4byte
 THPDec_804DE1D4 = .sdata2:0x804DE1D4; // type:object size:0x1 scope:local data:byte
 THPDec_804DE1D8 = .sdata2:0x804DE1D8; // type:object size:0x4 scope:local data:4byte
