+++ src/melee/ft/chara/ftKirby/ftkirbyspecialgamewatch.h
@@ -1,0 +1,28 @@
+#ifndef GALE01_10D074
+#define GALE01_10D074
+
+#include <placeholder.h>
+
+#include "ft/forward.h"
+
+/* 10D074 */ void ftKb_SpecialNGw_8010D074(Fighter_GObj* gobj);
+/* 10D0A8 */ void ftKb_SpecialNGw_8010D0A8(Fighter_GObj* gobj);
+/* 10D100 */ void fn_8010D100(Fighter_GObj* gobj);
+/* 10D130 */ void ftKb_SpecialNGw_8010D130(Fighter_GObj* gobj);
+/* 10D160 */ bool ftKb_SpecialNGw_8010D160(Fighter_GObj* gobj);
+/* 10D188 */ void ftKb_SpecialNGw_8010D188(Fighter_GObj* gobj);
+/* 10D204 */ void ftKb_SpecialNGw_8010D204(Fighter_GObj* gobj);
+/* 10D280 */ void ftKb_GwSpecialN_Anim(Fighter_GObj* gobj);
+/* 10D320 */ void ftKb_GwSpecialAirN_Anim(Fighter_GObj* gobj);
+/* 10D3C0 */ void ftKb_GwSpecialN_IASA(Fighter_GObj* gobj);
+/* 10D444 */ void ftKb_GwSpecialAirN_IASA(Fighter_GObj* gobj);
+/* 10D4C8 */ void ftKb_GwSpecialN_Phys(Fighter_GObj* gobj);
+/* 10D4E8 */ void ftKb_GwSpecialAirN_Phys(Fighter_GObj* gobj);
+/* 10D508 */ void ftKb_GwSpecialN_Coll(Fighter_GObj* gobj);
+/* 10D544 */ void ftKb_GwSpecialAirN_Coll(Fighter_GObj* gobj);
+/* 10D580 */ void ftKb_SpecialNGw_8010D580(Fighter_GObj* gobj);
+/* 10D5F0 */ void ftKb_SpecialNGw_8010D5F0(Fighter_GObj* gobj);
+/* 10D660 */ void ftKb_SpecialNGw_8010D660(Fighter_GObj* gobj, float);
+/* 10D6D0 */ void ftKb_SpecialNGw_8010D6D0(Fighter_GObj* gobj, float);
+
+#endif
