+++ src/melee/ft/chara/ftKirby/ftkirbyspecialmars.h
@@ -1,0 +1,46 @@
+#ifndef GALE01_10B1F4
+#define GALE01_10B1F4
+
+#include <placeholder.h>
+
+#include "ft/forward.h"
+
+/* 10B1F4 */ void fn_8010B1F4(Fighter_GObj* gobj);
+/* 10B2FC */ void ftKb_SpecialNMs_8010B2FC(Fighter_GObj* gobj);
+/* 10B4A0 */ void ftKb_SpecialNMs_8010B4A0(Fighter_GObj* gobj);
+/* 10B65C */ void ftKb_MsSpecialNStart_Anim(Fighter_GObj* gobj);
+/* 10B6C4 */ void ftKb_MsSpecialAirNStart_Anim(Fighter_GObj* gobj);
+/* 10B72C */ void ftKb_MsSpecialNStart_IASA(Fighter_GObj* gobj);
+/* 10B730 */ void ftKb_MsSpecialAirNStart_IASA(Fighter_GObj* gobj);
+/* 10B734 */ void ftKb_MsSpecialNStart_Phys(Fighter_GObj* gobj);
+/* 10B78C */ void ftKb_MsSpecialAirNStart_Phys(Fighter_GObj* gobj);
+/* 10B7F0 */ void ftKb_MsSpecialNStart_Coll(Fighter_GObj* gobj);
+/* 10B82C */ void ftKb_MsSpecialAirNStart_Coll(Fighter_GObj* gobj);
+/* 10B868 */ void ftKb_SpecialNMs_8010B868(Fighter_GObj* gobj);
+/* 10B8E0 */ void ftKb_SpecialNMs_8010B8E0(Fighter_GObj* gobj);
+/* 10B958 */ void ftKb_MsSpecialNLoop_Anim(Fighter_GObj* gobj);
+/* 10B9C0 */ void ftKb_MsSpecialAirNLoop_Anim(Fighter_GObj* gobj);
+/* 10BA28 */ void ftKb_MsSpecialNLoop_IASA(Fighter_GObj* gobj);
+/* 10BA60 */ void ftKb_MsSpecialAirNLoop_IASA(Fighter_GObj* gobj);
+/* 10BA98 */ void ftKb_MsSpecialNLoop_Phys(Fighter_GObj* gobj);
+/* 10BAB8 */ void ftKb_MsSpecialAirNLoop_Phys(Fighter_GObj* gobj);
+/* 10BAD8 */ void ftKb_MsSpecialNLoop_Coll(Fighter_GObj* gobj);
+/* 10BB14 */ void ftKb_MsSpecialAirNLoop_Coll(Fighter_GObj* gobj);
+/* 10BB50 */ void ftKb_SpecialNMs_8010BB50(Fighter_GObj* gobj);
+/* 10BBC8 */ void ftKb_SpecialNMs_8010BBC8(Fighter_GObj* gobj);
+/* 10BC40 */ void ftKb_SpecialNMs_8010BC40(Fighter_GObj* gobj);
+/* 10BC90 */ void ftKb_SpecialNMs_8010BC90(Fighter_GObj* gobj);
+/* 10BCE0 */ void ftKb_MsSpecialNEnd_Anim(Fighter_GObj* gobj);
+/* 10BDD8 */ void ftKb_MsSpecialAirNEnd_Anim(Fighter_GObj* gobj);
+/* 10BED0 */ void ftKb_MsSpecialNEnd_IASA(Fighter_GObj* gobj);
+/* 10BED4 */ void ftKb_MsSpecialAirNEnd_IASA(Fighter_GObj* gobj);
+/* 10BED8 */ void ftKb_MsSpecialNEnd_Phys(Fighter_GObj* gobj);
+/* 10BEF8 */ void ftKb_MsSpecialAirNEnd_Phys(Fighter_GObj* gobj);
+/* 10BF18 */ void ftKb_MsSpecialNEnd_Coll(Fighter_GObj* gobj);
+/* 10BF54 */ void ftKb_MsSpecialAirNEnd_Coll(Fighter_GObj* gobj);
+/* 10BF90 */ void ftKb_SpecialNPe_8010BF90(Fighter_GObj* gobj);
+/* 10C06C */ void ftKb_SpecialNPe_8010C06C(Fighter_GObj* gobj);
+/* 10C148 */ void ftKb_SpecialNPe_8010C148(Fighter_GObj* gobj);
+/* 10C1E8 */ void ftKb_SpecialNPe_8010C1E8(Fighter_GObj* gobj);
+
+#endif
