+++ src/melee/ft/chara/ftKirby/ftkirbyspecialpeach.h
@@ -1,0 +1,34 @@
+#ifndef GALE01_10C288
+#define GALE01_10C288
+
+#include "ft/forward.h"
+
+/* 10C3C0 */ void ftKb_SpecialNPe_8010C3C0(Fighter_GObj* gobj);
+/* 10C3F4 */ void ftKb_SpecialNPe_8010C3F4(Fighter_GObj* gobj);
+/* 10C47C */ void ftKb_SpecialNPe_8010C47C(Fighter_GObj* gobj);
+/* 10C4AC */ bool ftKb_SpecialNPe_8010C4AC(Fighter_GObj* gobj);
+/* 10C4D4 */ void ftKb_SpecialNPe_8010C4D4(Fighter_GObj* gobj);
+/* 10C560 */ void ftKb_SpecialNPe_8010C560(Fighter_GObj* gobj);
+/* 10C5FC */ void ftKb_PeSpecialLw_Anim(Fighter_GObj* gobj);
+/* 10C6B8 */ void ftKb_PeSpecialAirLw_Anim(Fighter_GObj* gobj);
+/* 10C774 */ void ftKb_PeSpecialLw_IASA(Fighter_GObj* gobj);
+/* 10C778 */ void ftKb_PeSpecialAirLw_IASA(Fighter_GObj* gobj);
+/* 10C77C */ void ftKb_PeSpecialLw_Phys(Fighter_GObj* gobj);
+/* 10C7B0 */ void ftKb_PeSpecialAirLw_Phys(Fighter_GObj* gobj);
+/* 10C860 */ void ftKb_PeSpecialLw_Coll(Fighter_GObj* gobj);
+/* 10C89C */ void ftKb_PeSpecialAirLw_Coll(Fighter_GObj* gobj);
+/* 10C8D8 */ void ftKb_SpecialNPe_8010C8D8(Fighter_GObj* gobj);
+/* 10C9CC */ void ftKb_SpecialNPe_8010C9CC(Fighter_GObj* gobj);
+/* 10CAB4 */ void ftKb_PeSpecialLwHit_Anim(Fighter_GObj* gobj);
+/* 10CB18 */ void ftKb_PeSpecialAirLwHit_Anim(Fighter_GObj* gobj);
+/* 10CB7C */ void ftKb_PeSpecialLwHit_IASA(Fighter_GObj* gobj);
+/* 10CB80 */ void ftKb_PeSpecialAirLwHit_IASA(Fighter_GObj* gobj);
+/* 10CB84 */ void ftKb_PeSpecialLwHit_Phys(Fighter_GObj* gobj);
+/* 10CBA4 */ void ftKb_PeSpecialAirLwHit_Phys(Fighter_GObj* gobj);
+/* 10CBF4 */ void ftKb_PeSpecialLwHit_Coll(Fighter_GObj* gobj);
+/* 10CC30 */ void ftKb_PeSpecialAirLwHit_Coll(Fighter_GObj* gobj);
+/* 10CC6C */ void ftKb_SpecialNGw_8010CC6C(Fighter_GObj* gobj);
+/* 10CCD4 */ void ftKb_SpecialNGw_8010CCD4(Fighter_GObj* gobj);
+/* 10CD44 */ void ftKb_SpecialNGw_8010CD44(Fighter_GObj* gobj);
+
+#endif
