+++ src/melee/ft/chara/ftKirby/ftkirbyyoshiegg.h
@@ -13,88 +13,5 @@
 /* 10B148 */ void fn_8010B148(Fighter_GObj* gobj);
 /* 10B16C */ void fn_8010B16C(Fighter_GObj* gobj);
 /* 10B1D4 */ void fn_8010B1D4(Fighter_GObj* gobj);
-/* 10B1F4 */ void fn_8010B1F4(Fighter_GObj* gobj);
-/* 10B2FC */ void ftKb_SpecialNMs_8010B2FC(Fighter_GObj* gobj);
-/* 10B4A0 */ void ftKb_SpecialNMs_8010B4A0(Fighter_GObj* gobj);
-/* 10B65C */ void ftKb_MsSpecialNStart_Anim(Fighter_GObj* gobj);
-/* 10B6C4 */ void ftKb_MsSpecialAirNStart_Anim(Fighter_GObj* gobj);
-/* 10B72C */ void ftKb_MsSpecialNStart_IASA(Fighter_GObj* gobj);
-/* 10B730 */ void ftKb_MsSpecialAirNStart_IASA(Fighter_GObj* gobj);
-/* 10B734 */ void ftKb_MsSpecialNStart_Phys(Fighter_GObj* gobj);
-/* 10B78C */ void ftKb_MsSpecialAirNStart_Phys(Fighter_GObj* gobj);
-/* 10B7F0 */ void ftKb_MsSpecialNStart_Coll(Fighter_GObj* gobj);
-/* 10B82C */ void ftKb_MsSpecialAirNStart_Coll(Fighter_GObj* gobj);
-/* 10B868 */ void ftKb_SpecialNMs_8010B868(Fighter_GObj* gobj);
-/* 10B8E0 */ void ftKb_SpecialNMs_8010B8E0(Fighter_GObj* gobj);
-/* 10B958 */ void ftKb_MsSpecialNLoop_Anim(Fighter_GObj* gobj);
-/* 10B9C0 */ void ftKb_MsSpecialAirNLoop_Anim(Fighter_GObj* gobj);
-/* 10BA28 */ void ftKb_MsSpecialNLoop_IASA(Fighter_GObj* gobj);
-/* 10BA60 */ void ftKb_MsSpecialAirNLoop_IASA(Fighter_GObj* gobj);
-/* 10BA98 */ void ftKb_MsSpecialNLoop_Phys(Fighter_GObj* gobj);
-/* 10BAB8 */ void ftKb_MsSpecialAirNLoop_Phys(Fighter_GObj* gobj);
-/* 10BAD8 */ void ftKb_MsSpecialNLoop_Coll(Fighter_GObj* gobj);
-/* 10BB14 */ void ftKb_MsSpecialAirNLoop_Coll(Fighter_GObj* gobj);
-/* 10BB50 */ void ftKb_SpecialNMs_8010BB50(Fighter_GObj* gobj);
-/* 10BBC8 */ void ftKb_SpecialNMs_8010BBC8(Fighter_GObj* gobj);
-/* 10BC40 */ void ftKb_SpecialNMs_8010BC40(Fighter_GObj* gobj);
-/* 10BC90 */ void ftKb_SpecialNMs_8010BC90(Fighter_GObj* gobj);
-/* 10BCE0 */ void ftKb_MsSpecialNEnd_Anim(Fighter_GObj* gobj);
-/* 10BDD8 */ void ftKb_MsSpecialAirNEnd_Anim(Fighter_GObj* gobj);
-/* 10BED0 */ void ftKb_MsSpecialNEnd_IASA(Fighter_GObj* gobj);
-/* 10BED4 */ void ftKb_MsSpecialAirNEnd_IASA(Fighter_GObj* gobj);
-/* 10BED8 */ void ftKb_MsSpecialNEnd_Phys(Fighter_GObj* gobj);
-/* 10BEF8 */ void ftKb_MsSpecialAirNEnd_Phys(Fighter_GObj* gobj);
-/* 10BF18 */ void ftKb_MsSpecialNEnd_Coll(Fighter_GObj* gobj);
-/* 10BF54 */ void ftKb_MsSpecialAirNEnd_Coll(Fighter_GObj* gobj);
-/* 10BF90 */ void ftKb_SpecialNPe_8010BF90(Fighter_GObj* gobj);
-/* 10C06C */ void ftKb_SpecialNPe_8010C06C(Fighter_GObj* gobj);
-/* 10C148 */ void ftKb_SpecialNPe_8010C148(Fighter_GObj* gobj);
-/* 10C1E8 */ void ftKb_SpecialNPe_8010C1E8(Fighter_GObj* gobj);
-/* 10C3C0 */ void ftKb_SpecialNPe_8010C3C0(Fighter_GObj* gobj);
-/* 10C3F4 */ void ftKb_SpecialNPe_8010C3F4(Fighter_GObj* gobj);
-/* 10C47C */ void ftKb_SpecialNPe_8010C47C(Fighter_GObj* gobj);
-/* 10C4AC */ bool ftKb_SpecialNPe_8010C4AC(Fighter_GObj* gobj);
-/* 10C4D4 */ void ftKb_SpecialNPe_8010C4D4(Fighter_GObj* gobj);
-/* 10C560 */ void ftKb_SpecialNPe_8010C560(Fighter_GObj* gobj);
-/* 10C5FC */ void ftKb_PeSpecialLw_Anim(Fighter_GObj* gobj);
-/* 10C6B8 */ void ftKb_PeSpecialAirLw_Anim(Fighter_GObj* gobj);
-/* 10C774 */ void ftKb_PeSpecialLw_IASA(Fighter_GObj* gobj);
-/* 10C778 */ void ftKb_PeSpecialAirLw_IASA(Fighter_GObj* gobj);
-/* 10C77C */ void ftKb_PeSpecialLw_Phys(Fighter_GObj* gobj);
-/* 10C7B0 */ void ftKb_PeSpecialAirLw_Phys(Fighter_GObj* gobj);
-/* 10C860 */ void ftKb_PeSpecialLw_Coll(Fighter_GObj* gobj);
-/* 10C89C */ void ftKb_PeSpecialAirLw_Coll(Fighter_GObj* gobj);
-/* 10C8D8 */ void ftKb_SpecialNPe_8010C8D8(Fighter_GObj* gobj);
-/* 10C9CC */ void ftKb_SpecialNPe_8010C9CC(Fighter_GObj* gobj);
-/* 10CAB4 */ void ftKb_PeSpecialLwHit_Anim(Fighter_GObj* gobj);
-/* 10CB18 */ void ftKb_PeSpecialAirLwHit_Anim(Fighter_GObj* gobj);
-/* 10CB7C */ void ftKb_PeSpecialLwHit_IASA(Fighter_GObj* gobj);
-/* 10CB80 */ void ftKb_PeSpecialAirLwHit_IASA(Fighter_GObj* gobj);
-/* 10CB84 */ void ftKb_PeSpecialLwHit_Phys(Fighter_GObj* gobj);
-/* 10CBA4 */ void ftKb_PeSpecialAirLwHit_Phys(Fighter_GObj* gobj);
-/* 10CBF4 */ void ftKb_PeSpecialLwHit_Coll(Fighter_GObj* gobj);
-/* 10CC30 */ void ftKb_PeSpecialAirLwHit_Coll(Fighter_GObj* gobj);
-/* 10CC6C */ void ftKb_SpecialNGw_8010CC6C(Fighter_GObj* gobj);
-/* 10CCD4 */ void ftKb_SpecialNGw_8010CCD4(Fighter_GObj* gobj);
-/* 10CD44 */ void ftKb_SpecialNGw_8010CD44(Fighter_GObj* gobj);
-/* 10D074 */ void ftKb_SpecialNGw_8010D074(Fighter_GObj* gobj);
-/* 10D0A8 */ void ftKb_SpecialNGw_8010D0A8(Fighter_GObj* gobj);
-/* 10D100 */ void fn_8010D100(Fighter_GObj* gobj);
-/* 10D130 */ void ftKb_SpecialNGw_8010D130(Fighter_GObj* gobj);
-/* 10D160 */ bool ftKb_SpecialNGw_8010D160(Fighter_GObj* gobj);
-/* 10D188 */ void ftKb_SpecialNGw_8010D188(Fighter_GObj* gobj);
-/* 10D204 */ void ftKb_SpecialNGw_8010D204(Fighter_GObj* gobj);
-/* 10D280 */ void ftKb_GwSpecialN_Anim(Fighter_GObj* gobj);
-/* 10D320 */ void ftKb_GwSpecialAirN_Anim(Fighter_GObj* gobj);
-/* 10D3C0 */ void ftKb_GwSpecialN_IASA(Fighter_GObj* gobj);
-/* 10D444 */ void ftKb_GwSpecialAirN_IASA(Fighter_GObj* gobj);
-/* 10D4C8 */ void ftKb_GwSpecialN_Phys(Fighter_GObj* gobj);
-/* 10D4E8 */ void ftKb_GwSpecialAirN_Phys(Fighter_GObj* gobj);
-/* 10D508 */ void ftKb_GwSpecialN_Coll(Fighter_GObj* gobj);
-/* 10D544 */ void ftKb_GwSpecialAirN_Coll(Fighter_GObj* gobj);
-/* 10D580 */ void ftKb_SpecialNGw_8010D580(Fighter_GObj* gobj);
-/* 10D5F0 */ void ftKb_SpecialNGw_8010D5F0(Fighter_GObj* gobj);
-/* 10D660 */ void ftKb_SpecialNGw_8010D660(Fighter_GObj* gobj, float);
-/* 10D6D0 */ void ftKb_SpecialNGw_8010D6D0(Fighter_GObj* gobj, float);
 
 #endif
