+++ src/melee/it/items/itkirbygamewatchchefpan.c
@@ -1,6 +1,6 @@
 #include "itkirbygamewatchchefpan.h"
 
-#include "ftKirby/ftkirbyyoshiegg.h"
+#include "ftKirby/ftkirbyspecialgamewatch.h"
 #include "it/itzako.h"
 
 #include <melee/db/db.h>
@@ -11,12 +11,7 @@
 #include <melee/it/types.h>
 
 ItemStateTable it_803F7958[] = {
-    {
-        -1,
-        itKirbygamewatchchefpan_UnkMotion0_Anim,
-        NULL,
-        NULL,
-    },
+    { -1, itKirbygamewatchchefpan_UnkMotion0_Anim, NULL, NULL },
 };
 
 Item_GObj* it_802C74D8(HSD_GObj* parent, Vec3* pos, Fighter_Part part,
